complete refactor.
